### 2020-01-30 #14 @khinsen
[Rp] Structural flexibility in proteins: impact of the crystal environment
**Original article:** [Hinsen, K (2008) Structural flexibility in proteins: impact of the crystal environment. Bioinformatics 24:521.](https://doi.org/10.1093/bioinformatics/btm625) ([Open Access version](https://academic.oup.com/bioinformatics/article-pdf/24/4/521/16884353/btm625.pdf))

**Scientific domain:**  Biophysics
**Suggested editor:**  @delsuc
