### 2020-02-07 #14 @fxcoudert
@delsuc Here is my review

## Paper

I quite appreciated the non-nonsense writing of the paper, which is good for a reproducibility study to show the pitfalls and issues, even when they turn out not to be crucial.

- Regarding the numpy evolutions, which I think are quite important to the community at large, maybe this could be better characterised: what is the numpy policy on breaking changes, deprecation process, and use of semantic versioning? Maybe links to this information could be added. (But that opens another 10-year challenge: if a paper has web links, how many will still be active 10 years later?)
- The instructions for reproduction are rather clear, and the recommendation to use Guix is interesting (I know of it, and want to try it myself but haven't found the time yet).
- The main point I think could be discussed is the use of Python 2, given the EOL was earlier this year (2020-01-01), and it is being removed from software distributions. How long does the author suggest it will be feasible to install Python 2 on modern hardware? Or without a dedicated OS like Guix? I would be interested in comments on that, as I think it's a worrying topic.
- Grace is still working fine, I am using it daily :)
- I do not understand why there are two folders named `updated-scripts-from-obsolete-web-server` and `scripts-from-suplementary-material-to-original-paper`, and they are not documented in the paper. One of them has an extra script (`calculate_crystal_fluctuations_small_memory.py`) which is not mentioned anywhere else either. _Edit: I see that the notes file mentions those. Probably best to mention that in the paper itself, though._

## Typo

- “there is longer any point in such a change” (page 1)
