### 2020-02-07 #14 @delsuc
Thank you very much @fxcoudert !
@khinsen can you improve the document as mentionned ?
