+++ article/content.tex
@@ -4,7 +4,7 @@ The article that is the subject of this reproduction attempt was my first public
 
 \section*{Historical context}
 
-The work described in the original article was performed in summer 2007. The scripts make heavy use of two libraries of which I am the principal author: the Molecular Modeling Toolkit (MMTK) \cite{Hinsenmolecularmodelingtoolkit2000}, then at version 2.5, and ScientificPython, then at version 2.7. These two libraries, first published in 1997, are among the oldest scientific computing packages in the SciPy ecosystem. They were initially written on the basis of Numerical Python \cite{DuboisNumericalPython1996}, the original array package for Python that was published in 1995. With the transition of the ecosystem to its successor NumPy \cite{OliphantguideNumPy2006}, initially released in 2006, I ported MMTK and ScientificPython using NumPy's compatibility module called \texttt{oldnumeric}. This combination was used for the original work, but unfortunately I did not write down the exact version of NumPy. My instructions recommend ``Numerical Python 23.8.2 or NumPy 1.x'', which turned out to be overly optimistic: NumPy is still in the 1.x version range, but frequent breaking changes make it impossible to run my code with recent NumPy releases. In fact, with version~1.9 NumPy removed the \texttt{oldnumeric} interface, breaking all of my molecular simulation code. I had envisaged to port the code to the official NumPy API, but decided not to do so because of the high risk of introducing errors. There are in fact several subtle changes in the API, such as using the transpose of a matrix compared to the original Numeric API, which are easy to overlook because incorrect use yields an incorrect result but no error message. With the end of support for Python~2, there is longer any point in such a change, as it makes little difference if my code depends on one or two unmaintained software packages.
+The work described in the original article was performed in summer 2007. The scripts make heavy use of two libraries of which I am the principal author: the Molecular Modeling Toolkit (MMTK) \cite{Hinsenmolecularmodelingtoolkit2000}, then at version 2.5, and ScientificPython, then at version 2.7. These two libraries, first published in 1997, are among the oldest scientific computing packages in the SciPy ecosystem. They were initially written on the basis of Numerical Python \cite{DuboisNumericalPython1996}, the original array package for Python that was published in 1995. With the transition of the ecosystem to its successor NumPy \cite{OliphantguideNumPy2006}, initially released in 2006, I ported MMTK and ScientificPython using NumPy's compatibility module called \texttt{oldnumeric}. This combination was used for the original work, but unfortunately I did not write down the exact version of NumPy. My instructions recommend ``Numerical Python 23.8.2 or NumPy 1.x'', which turned out to be overly optimistic: NumPy is still in the 1.x version range, but frequent breaking changes make it impossible to run my code with recent NumPy releases. In fact, with version~1.9 NumPy removed the \texttt{oldnumeric} interface, breaking all of my molecular simulation code. I had envisaged to port the code to the official NumPy API, but decided not to do so because of the high risk of introducing errors. There are in fact several subtle changes in the API, such as using the transpose of a matrix compared to the original Numeric API, which are easy to overlook because incorrect use yields an incorrect result but no error message. With the end of support for Python~2, there is no longer any point in such a change, as it makes little difference if my code depends on one or two unmaintained software packages.
 
 The remaining dependencies turned out to be less critical. netCDF, which my scripts use for data storage, was at version 3.4, but any more recent version can be substituted. Python itself, then at version 2.5, caused no problems either up to release 2.7. Python~3 is from my point of view a different language, the two major changes for scientific computing being a very different C API layer and a different definition of division. Porting would involve a significant amount of work because of the large number of extension modules in MMTK, most of which are C code hand-written for the C~API of Python 1.4, long before Pyrex \cite{EwingPyrex} and then its fork Cython \cite{BehnelCythonBestBoth2011} introduced more convenient ways to write extension modules.
 
