+++ article/content.tex
@@ -32,7 +32,7 @@ In addition to the fix described above, I modified my scripts for convenience, m
 
 \subsection*{Reproducing the figures}
 
-As explained in the introduction, my goal with publishing the code of my computations was to enable reuse, not reproducibility. The code therefore does not reproduce the full figures, which need to be re-generated by hand from the numerical output. I have limited myself to producing figures that are similar enough to the originals to convince the reader that the results have been reproduced correctly. I also used different plotting software, Gnuplot \cite{WilliamsGnuplot2020}, in replacement of the originally used Grace \cite{GraceDevelopmentTeamGrace2008} that seems to be no longer supported.
+As explained in the introduction, my goal with publishing the code of my computations was to enable reuse, not reproducibility. The code therefore does not reproduce the full figures, which need to be re-generated by hand from the numerical output. I have limited myself to producing figures that are similar enough to the originals to convince the reader that the results have been reproduced correctly. I also used different plotting software, Gnuplot \cite{WilliamsGnuplot2020}, in replacement of the originally used Grace \cite{GraceDevelopmentTeamGrace2008} which has not yet been packaged for Guix.
 
 Figure~1 of the original publication does not contain any computed data and has therefore not been reproduced. The data it shows comes directly from the Protein Data Bank.
 
