### 2020-02-20 a306649
Revision: explain the two versions of the scripts in the repository
