### 2020-02-20 70d6720
Revision: discussion of Python 2 and NumPy
