+++ article/bibliography.bib
@@ -1,4 +1,12 @@
 
+@misc{AnacondaInc.Anacondasoftwaredistribution2020,
+  title = {Anaconda Software Distribution},
+  author = {{Anaconda Inc.}},
+  year = {2020},
+  month = feb,
+  howpublished = {\\ https://web.archive.org/web/20200220005300/https://www.anaconda.com/}
+}
+
 @book{Arthurnaturetechnologywhat2009,
   ids = {Arthurnaturetechnologywhat2009a},
   title = {The Nature of Technology: What It Is and How It Evolves},
@@ -8,7 +16,6 @@
   publisher = {{Free Press}},
   address = {{New York, NY}},
   isbn = {978-1-4165-4405-0 978-1-4391-6578-2},
-  language = {eng},
   note = {OCLC: 695688782}
 }
 
@@ -50,20 +57,51 @@
   copyright = {All rights reserved},
   file = {/home/hinsen/Zotero/storage/HFMHTEXP/Dubois et al. - 1996 - Numerical Python.pdf},
   journal = {Computers in Physics},
-  language = {en},
   number = {3}
 }
 
-@misc{EwingPyrex,
+@misc{EwingPyrex2010,
   title = {Pyrex},
-  author = {Ewing, Greg}
+  author = {Ewing, Greg},
+  year = {2010},
+  howpublished = {\\ https://web.archive.org/web/20200216002742/https://www.cosc.canterbury.ac.nz/greg.ewing/python/Pyrex/}
+}
+
+@misc{GommersNEP23Backwards2018,
+  title = {{{NEP}} 23 \textemdash{} {{Backwards}} Compatibility and Deprecation Policy},
+  author = {Gommers, Ralf},
+  year = {2018},
+  howpublished = {\\ https://web.archive.org/web/20191007231112/https://numpy.org/neps/nep-0023-backwards-compatibility.html},
+  journal = {NumPy documentation}
 }
 
 @misc{GraceDevelopmentTeamGrace2008,
   title = {Grace},
-  author = {Grace Development Team},
+  author = {{Grace Development Team}},
   year = {2008},
-  howpublished = {http://plasma-gate.weizmann.ac.il/Grace/}
+  howpublished = {\\ http://plasma-gate.weizmann.ac.il/Grace/}
+}
+
+@inproceedings{GreenfieldnumarrayNewScientific2003,
+  title = {Numarray: {{A New Scientific Array Package}} for {{Python}}},
+  booktitle = {{{PyCon DC}}},
+  author = {Greenfield, Perry and Miller, Jay Todd and Hsu, Jin-chung and White, Richard L},
+  year = {2003},
+  file = {/home/hinsen/Zotero/storage/6BJBM37I/Greenfield et al. - numarray A New Scientific Array Package for Pytho.pdf}
+}
+
+@article{HinsenDealingSoftwareCollapse2019,
+  title = {Dealing {{With Software Collapse}}},
+  author = {Hinsen, Konrad},
+  year = {2019},
+  month = may,
+  volume = {21},
+  pages = {104--108},
+  issn = {1521-9615, 1558-366X},
+  doi = {10.1109/MCSE.2019.2900945},
+  file = {/home/hinsen/Zotero/storage/ER3Q9B5U/Hinsen - 2019 - Dealing With Software Collapse.pdf},
+  journal = {Comput. Sci. Eng.},
+  number = {3}
 }
 
 @article{Hinsenmolecularmodelingtoolkit2000,
@@ -94,6 +132,22 @@
   number = {4}
 }
 
+@misc{NumPydevelopersandusersRequestusesemantic,
+  title = {Request: Use Semantic Versioning},
[…]
