+++ article/content.tex
@@ -14,7 +14,7 @@ The \href{https://github.com/khinsen/rescience-ten-year-challenge-paper-3}{GitHu
 
 \subsection*{Finding the code and the input data}
 
-The project-specific scripts are still available for download from the journal's Web site. A copy can be found in the directory \texttt{scripts-from-suplementary-material-to-original-paper} of the code repository for this article. The directory \texttt{updated-scripts-from-obsolete-web-server} contains updated versions of the scripts, added for completeness. They were available from my laboratory's Web site for a few years, and contain minor improvements that are not relevant for this reproduction attempt, see the file \texttt{notes.org} for details. In the following, I use the original scripts.
+The project-specific scripts are still available for download from the journal's Web site. A copy can be found in the directory \scalebox{.7}[1.0]{\texttt{scripts-from-suplementary-material-to-original-paper}} of the code repository for this article. The directory \scalebox{.7}[1.0]{\texttt{updated-scripts-from-obsolete-web-server}} contains updated versions of the scripts, added for completeness. They were available from my laboratory's Web site for a few years, and contain minor improvements that are not relevant for this reproduction attempt, see the file \texttt{notes.org} for details. In the following, I use the original scripts.
 
 All the dependencies have been available in public version-controlled repositories for many years. Obtaining the published code is therefore not a problem. The additional scripts that generated the original plots were not published as explained in the introduction. They have not survived two changes of computers. I still have backups that should contain them, but I have not been able to find a working reader for the DAT/DDS tapes on which these backups are stored. Another backup on a CD-ROM turned out to be unreadable.
 
