+++ article/content.tex
@@ -94,8 +94,6 @@ The obvious lesson for the future to draw from this exercise is the importance o
 
 The final lesson would have to be drawn by a wider community of scientific software developers: breaking changes in widely used infrastructure code such as NumPy can cause a lot of damage in terms of lost reproducibility that may be difficult to diagnose and fix for someone else than the original authors. The open question that the scientific community has to figure out is where to place reproducibility on our scale of values, and for which time spans we consider it desirable. Structural biology is a methodologically mature domain of research, in which the main source of progress is not disruptive new methods, but incremental improvements of existing methods in the course of ongoing applications to biologically relevant systems. This is in fact the norm in science and technology \cite{Arthurnaturetechnologywhat2009}. In this context, my methodological work published in~2008 is by no means outdated. The last time I have referred to it myself (in so-far unpublished work) was in 2018. Reproducibility lifetimes of just a few years, corresponding to the current habits in the scientific Python ecosystem, are therefore problematic.
 
-% NumPy's predecessor \texttt{Numeric} \cite{DuboisNumericalPython1996} used a semantic versioning scheme, and progressed from its first stable version~1 (released in March 1998) to version~24 in its decade of active development. Nevertheless, the changes in the major version number mostly reflected important new features. Backwards incompatible changes were, according to my memories, rare \footnote{Unfortunately the source code, and thus the release notes, for Numerical Python versions before~16 (first one to be published on SourceForge) is probably lost forever.} Today's NumPy started out as a merger of Numeric with an incompatible fork called \texttt{numarray}. 
-
 \vspace{3mm}
 
 The two pieces of software whose evolution has most impacted this work, Python and NumPy, deserve a more detailed discussion because of their importance in today's scientific computing. In the 12~years since the publication of my original paper, both have changed in a way that breaks compatibility with earlier version, causing software collapse \cite{HinsenDealingSoftwareCollapse2019} that has affected not just my two scripts, but much scientific software. However, the stories of the two pieces of software are also very different.
