### 2020-02-21 #14 @khinsen
@delsuc @fxcoudert The revision is ready!

BIg change: I have added a section at the end of the Conclusions that discusses the evolution of Python 2 and NumPy in the context of computational science.

Small changes: all the other remarks in the review.

In preparation for the ten-year link rot challenge, I only refer to archived copies from the [Internet Archive](https://archive.org/index.php). Let's hope it will stay around for ten more years at least!
