### 2020-03-15 #14 @delsuc
I believe this is ready for acceptation and publication.
@fxcoudert is that ok with you ?
