### 2020-03-19 #14 @delsuc
Sorry @khinsen for the delay - the period is a bit complex to deal with... :-/
Of course the manuscript is accepted to be published is ReScience.
I am not sure what is the procedure at this level...
