+++ article/metadata.yaml
@@ -96,7 +96,7 @@ contributors:
 # This information will be provided by the editor
 dates:
   - received: January 30, 2020
-  - accepted: 
+  - accepted: March 19, 2020
   - published: 
 
 # This information will be provided by the editor
