### 2020-03-19 #14 @khinsen
Thanks @delsuc !

The procedure for actually publishing the article is described [here](https://github.com/rescience/articles). Roughly, the steps are
- assigning an article number
- reserving a DOI on Zenodo
- produce a new PDF with the DOI (that's a job for me)
- publish the PDF on Zenodo
- add the article to the ReScience table of contents.

Since this is purely technical, with no decisions involved, I can actually do all of this myself if you prefer. I see no potential conflict of interest. But if you want to practice the job which allows certain commercial publishers to make lots of money, go ahead and ask @rougier or myself for help if you get stuck (the current procedure is still rather new).
