### 2020-06-09 #14 @khinsen
This article has been published (http://doi.org/10.5281/zenodo.3886447) and will soon be listed on the ReScience Web site.
